完善用户系统，增加手机号登录，注册，找回密码；

+++ YKCenterDemo/YKRemoteViewController.m
@@ -37,7 +37,7 @@
     __weak __typeof(self)weakSelf = self;
     [YKCenterSDK learnCodeWithYKCId:[[YKCenterCommon sharedInstance] currentYKCId]
                          completion:^(BOOL result, NSString * _Nullable code) {
-                             NSLog(@"result:%ld, code:%@", result, code);
+                             NSLog(@"result:%ld, code:%@", (long)result, code);
                              if (result) {
                                  // 使用举例：修改第一个key的码值
                                  YKRemoteDeviceKey *key = weakSelf.remote.keys[0];
